Dockerfiles/docker-api_service/requirements.txt
flask==2.0.0
Flask-SQLAlchemy==2.5.1
Pyjwt==2.1.0